Refactor backend and frontend code
